♻️ refactor: Convert supabase functions to endpoint classes, consolidate types into one file, reorganise files, fix imports
