+++ app/(auth)/registration.tsx
@@ -1,10 +1,9 @@
 import { useAuth } from '@/providers/AuthProvider';
 import { showToast } from '@/utils/helpers';
-import { saveImage } from '@/utils/users';
+import * as ImagePicker from 'expo-image-picker';
 import { useRouter } from 'expo-router';
 import { useState } from 'react';
-import { Text, TextInput, TouchableOpacity, View, Image } from 'react-native';
-import * as ImagePicker from 'expo-image-picker';
+import { Image, Text, TextInput, TouchableOpacity, View } from 'react-native';
 
 export default function RegistrationScreen() {
   const router = useRouter();
