+++ components/OwnProfileHeader.tsx
@@ -1,13 +1,6 @@
-import { TProfile } from '@/utils/posts';
 import { useRouter } from 'expo-router';
 import React from 'react';
-import {
-  Image,
-  ImageSourcePropType,
-  Text,
-  TouchableOpacity,
-  View,
-} from 'react-native';
+import { Image, Text, TouchableOpacity, View } from 'react-native';
 
 interface IProfileHeader {
   postCount: number;
